changes from synchro
