+++ opencti-platform/opencti-front/src/components/ThemeLight.ts
@@ -34,6 +34,9 @@ const ThemeLight = (
       main: '#f44336',
       dark: '#c62828',
     },
+    warn: {
+      main: '#ffa726',
+    },
     dangerZone: { main: '#f6685e', light: '#fbc2be', dark: '#D1584F', contrastText: 'black', text: { primary: '#D1584F' } },
     success: { main: '#03a847' },
     primary: { main: primary || THEME_LIGHT_DEFAULT_PRIMARY },
