+++ opencti-platform/opencti-front/src/private/components/settings/sub_types/fintel_templates/FintelTemplatePopover.tsx
@@ -1,43 +1,18 @@
 import MoreVert from '@mui/icons-material/MoreVert';
 import React, { UIEvent, useState } from 'react';
 import { MenuItem, Menu, PopoverProps, IconButton } from '@mui/material';
-import { graphql } from 'react-relay';
-import DialogContent from '@mui/material/DialogContent';
-import DialogContentText from '@mui/material/DialogContentText';
-import DialogActions from '@mui/material/DialogActions';
-import Button from '@mui/material/Button';
-import Dialog from '@mui/material/Dialog';
-import { FintelTemplatePopoverDeleteMutation } from './__generated__/FintelTemplatePopoverDeleteMutation.graphql';
-import Transition from '../../../../../components/Transition';
-import useDeletion from '../../../../../utils/hooks/useDeletion';
-import useApiMutation from '../../../../../utils/hooks/useApiMutation';
 import { useFormatter } from '../../../../../components/i18n';
 import stopEvent from '../../../../../utils/domEvent';
-import { deleteNodeFromEdge } from '../../../../../utils/store';
-
-const fintelTemplatePopoverDeleteMutation = graphql`
-  mutation FintelTemplatePopoverDeleteMutation($id: ID!) {
-    fintelTemplateDelete(id: $id)
-  }
-`;
 
 interface FintelTemplatePopoverProps {
-  entitySettingId: string
-  templateId: string
   onUpdate: () => void
-  onDeleteComplete?: () => void
 }
 
 const FintelTemplatePopover = ({
-  entitySettingId,
-  templateId,
   onUpdate,
-  onDeleteComplete,
 }: FintelTemplatePopoverProps) => {
   const { t_i18n } = useFormatter();
   const [anchorEl, setAnchorEl] = useState<PopoverProps['anchorEl']>();
-
-  const [commitDeleteMutation] = useApiMutation<FintelTemplatePopoverDeleteMutation>(fintelTemplatePopoverDeleteMutation);
 
   const onOpenMenu = (e: UIEvent) => {
     stopEvent(e);
@@ -47,50 +22,6 @@ const FintelTemplatePopover = ({
   const onCloseMenu = (e: UIEvent) => {
     stopEvent(e);
     setAnchorEl(undefined);
-  };
-
-  const {
-    deleting,
-    handleOpenDelete,
-    displayDelete,
-    handleCloseDelete,
-    setDeleting,
-  } = useDeletion({});
-
-  const onAskDelete = (e: UIEvent) => {
-    stopEvent(e);
-    handleOpenDelete();
-    onCloseMenu(e);
-  };
-
-  const onCloseDelete = (e: UIEvent) => {
-    stopEvent(e);
-    handleCloseDelete();
-  };
-
-  const onDelete = (e: UIEvent) => {
-    stopEvent(e);
-    setDeleting(true);
-    commitDeleteMutation({
-      variables: { id: templateId },
-      updater: (store) => {
-        deleteNodeFromEdge(
-          store,
-          'fintelTemplates',
-          entitySettingId,
-          templateId,
-        );
-      },
-      onCompleted: () => {
-        setDeleting(false);
-        onCloseDelete(e);
-        onDeleteComplete?.();
-      },
-      onError: () => {
-        setDeleting(false);
-        onCloseDelete(e);
-      },
-    });
   };
 
   const update = (e: UIEvent) => {
@@ -107,30 +38,7 @@ const FintelTemplatePopover = ({
 
       <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} onClose={onCloseMenu}>
         <MenuItem onClick={update}>{t_i18n('Update')}</MenuItem>
-        <MenuItem onClick={onAskDelete}>{t_i18n('Delete')}</MenuItem>
       </Menu>
-
-      <Dialog
-        PaperProps={{ elevation: 1 }}
-        open={displayDelete}
-        keepMounted={true}
-        TransitionComponent={Transition}
-        onClose={onCloseDelete}
-      >
-        <DialogContent>
-          <DialogContentText>
-            {t_i18n('Do you want to delete this FINTEL template?')}
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={onCloseDelete} disabled={deleting}>
-            {t_i18n('Cancel')}
-          </Button>
-          <Button color="secondary" onClick={onDelete} disabled={deleting}>
-            {t_i18n('Delete')}
-          </Button>
-        </DialogActions>
-      </Dialog>
     </>
   );
 };
